third_party/blink/web_tests/fast/css/getComputedStyle/counterIncrement-without-counter-expected.txt

Test that the initial value of 'counter-increment' and 'counter-reset' is 'none'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.counterIncrement is "none"
PASS style.counterReset is "none"
PASS successfullyParsed is true

TEST COMPLETE
